+++ chat-service/src/main/java/com/ovengers/chatservice/mongodb/controller/MessageController.java
@@ -1,9 +1,11 @@
 package com.ovengers.chatservice.mongodb.controller;
 
+import com.ovengers.chatservice.mongodb.document.Message;
 import com.ovengers.chatservice.mongodb.dto.MessageDto;
 import com.ovengers.chatservice.mongodb.service.MessageService;
 import io.swagger.v3.oas.annotations.tags.Tag;
 import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 import reactor.core.publisher.Flux;
@@ -22,5 +24,30 @@ public class MessageController {
     public Mono<ResponseEntity<List<MessageDto>>> find(@PathVariable Long chatRoomId) {
         Flux<MessageDto> response = messageService.findMessages(chatRoomId);
         return response.collectList().map(ResponseEntity::ok);
+    }
+
+    @GetMapping("/{chatRoomId}/getMessages")
+    public Flux<Message> getMessageByChatRoomId(@PathVariable Long chatRoomId) {
+        return messageService.getMessageByChatRoomId(chatRoomId);
+    }
+
+    @PostMapping("/createMessage")
+    @ResponseStatus(HttpStatus.CREATED)
+    public Mono<Message> createMessage(@RequestBody Message message) {
+        return messageService.createMessage(message);
+    }
+
+    @PutMapping("/{messageId}/updateMessage")
+    public Mono<ResponseEntity<Message>> updateUser(@PathVariable String  messageId, @RequestBody Message message) {
+        return messageService.updateMessage(messageId, message)
+                .map(updatedMessage -> ResponseEntity.ok(updatedMessage))
+                .defaultIfEmpty(ResponseEntity.notFound().build());
+    }
+
+    @DeleteMapping("/{messageId}/deleteMessage")
+    public Mono<ResponseEntity<Void>> deleteUser(@PathVariable String messageId) {
+        return messageService.deleteMessage(messageId)
+                .then(Mono.just(ResponseEntity.ok().<Void>build()))
+                .defaultIfEmpty(ResponseEntity.notFound().build());
     }
 }
